Jenkinsfile

+++ Jenkinsfile
@@ -12,9 +12,18 @@ pipeline {
 
         stage('Install Dependencies') {
             steps {
-                sh 'sudo apt update'
-                sh 'sudo apt install nodejs'
-                sh 'sudo apt install npm'
+                script {
+                    if (fileExists('node_modules')){
+                        echo 'Dependencies already installed, skipping installation...'
+
+                    }else {
+                        echo 'installing dependencies....'
+                        sh 'sudo apt update'
+                        sh 'sudo apt install nodejs'
+                        sh 'sudo apt install npm'
+                    }
+                }
+                
                 echo 'successfully install node'
             }
         }
